requirements.txt

Django==5.1.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.1
dj-database-url==2.2.0
gunicorn==23.0.0
Markdown==3.7
markdown-it-py==3.0.0
django-cloudinary-storage==0.3.0
cloudinary==1.41.0
pillow==10.2.0
django-ckeditor==6.7.1
psycopg2-binary==2.9.9
django-allauth==65.0.2
PyJWT==2.8.0
cryptography==42.0.8
python-decouple==3.8
sseclient-py==1.8.0
MarkupSafe==2.1.5